refactor: move error handlers and responses; make global handler convey specific error messages  (#1)
