+++ src/main/java/io/hhplus/tdd/point/common/ErrorResponse.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd;
+package io.hhplus.tdd.point.common;
 
 public record ErrorResponse(
         String code,
